requirements.txt
flask
numpy
gunicorn
mediapipe
opencv-python
tensorflow-cpu
requests